Create Strassen matrix multiplication
Strassen algorithm for matrix multiplication is a faster way to multiply matrices. For a 2-by-2 matrix multiplied with another 2-by-2 matrix it reduces number of steps to 7 from the original 8 - this when coupled with a Divide-And-Conquer technique reduces the number of steps by one at each step resulting in log(7) nestings.

Note: Assumes the order of matrices to be power of 2. In case of other orders padding can be used to achieve power of 2 order.